refactor(config): extract auth into its own module and use a real config/mod.rs
