+++ src/commands/config/delete.rs
@@ -4,8 +4,8 @@ use clap::Command;
 
 use crate::cli::problem_json::http_error;
 use crate::cli::style;
+use crate::config::auth::load_auth_config;
 use crate::config::config::Config;
-use crate::config::config::load_auth_config;
 use crate::exit_code;
 
 pub(crate) fn command_config() -> Command {
