+++ src/commands/deployment/list.rs
@@ -2,8 +2,8 @@ use crate::api::dto::deployment::DeploymentOutput;
 use crate::cli::output::{output_arg, output_format};
 use crate::cli::problem_json::http_error_list;
 use crate::cli::style;
+use crate::config::auth::load_auth_config;
 use crate::config::config::Config;
-use crate::config::config::load_auth_config;
 use crate::exit_code;
 use clap::Arg;
 use clap::ArgMatches;
