+++ src/commands/deployment/logs.rs
@@ -1,5 +1,5 @@
+use crate::config::auth::load_auth_config;
 use crate::config::config::Config;
-use crate::config::config::load_auth_config;
 use crate::exit_code;
 use crate::runtime::lifecycle_trait::Log;
 use clap::{Arg, ArgAction, ArgMatches, Command};
