+++ src/commands/secret/delete.rs
@@ -6,8 +6,8 @@ use std::io::{self, Write};
 
 use crate::cli::problem_json::http_error;
 use crate::cli::style;
+use crate::config::auth::load_auth_config;
 use crate::config::config::Config;
-use crate::config::config::load_auth_config;
 use crate::exit_code;
 
 pub(crate) fn command_config() -> Command {
