+++ src/commands/user/create.rs
@@ -4,7 +4,8 @@ use clap::Command;
 use serde_json::json;
 
 use crate::cli::problem_json::render_response_error;
-use crate::config::config::{Config, load_auth_config};
+use crate::config::auth::load_auth_config;
+use crate::config::config::Config;
 use crate::exit_code;
 
 pub(crate) fn command_config() -> Command {
