+++ src/commands/user/update.rs
@@ -1,8 +1,8 @@
 use crate::api::dto::user::UserOutput;
 use crate::cli::problem_json::{http_error, render_response_error};
 use crate::cli::style;
+use crate::config::auth::load_auth_config;
 use crate::config::config::Config;
-use crate::config::config::load_auth_config;
 use crate::exit_code;
 use clap::Arg;
 use clap::ArgMatches;
