+++ src/commands/webhook/inspect.rs
@@ -1,6 +1,7 @@
 use crate::cli::problem_json::http_error;
 use crate::cli::style;
-use crate::config::config::{Config, load_auth_config};
+use crate::config::auth::load_auth_config;
+use crate::config::config::Config;
 use crate::exit_code;
 use clap::{Arg, ArgMatches, Command};
 use cli_table::{Table, WithTitle};
